pyproject.toml
[project]
name = "ckanext-statistics"
version = "2.0.15"
description = "A CKAN extension for accessing instance statistics."
readme = "README.md"
requires-python = ">=3.6"
license = { text = "GPL-3.0-or-later" }
authors = [
    { name = "Natural History Museum", email = "data@nhm.ac.uk" }
]
keywords = ["CKAN", "data", "statistics"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8"
]
dependencies = [
    "ckantools>=0.3.0",
    "requests",
    "tqdm>=4.55.1",
    "ckanext-ckanpackager>=2.2.0",
    "ckanext-versioned-datastore>=3.7.0"
]

[project.optional-dependencies]
test = [
    "mock",
    "pytest>=4.6.5",
    "pytest-cov>=2.7.1",
    "coveralls"
]

[project.urls]
repository = "https://github.com/NaturalHistoryMuseum/ckanext-statistics"
changelog = "https://github.com/NaturalHistoryMuseum/ckanext-statistics/blob/main/CHANGELOG.md"

[project.entry-points."ckan.plugins"]
statistics = "ckanext.statistics.plugin:StatisticsPlugin"


[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"


[tool]
[tool.setuptools]
zip-safe = false

[tool.setuptools.packages.find]
exclude = ["tests", "docs"]

[tool.commitizen]
name = "cz_nhm"
version = "2.0.15"
tag_format = "v$version"
update_changelog_on_bump = true
changelog_incremental = true
version_files = [
    "pyproject.toml:version"
]

[tool.black]
line-length = 88
skip_string_normalization = true

[tool.pylint]
max-line-length = 88
disable = ["C0114", "R0903"]

[tool.docformatter]
wrap-summaries = 88
wrap-descriptions = 88
pre-summary-newline = true
make-summary-multi-line = true
